user in activities

+++ api/routers/activity_router.py
@@ -14,6 +14,8 @@ from api.models.activity_models import (
     ActivityMediaUpdate,
 )
 from api.security.auth import verify_jwt_token
+from api.security.auth import get_current_user
+from api.models.user_models import Users
 
 # This file contains API endpoints related to Activities
 
@@ -44,8 +46,10 @@ def get_activity(session: SessionDep, activity_id: int):
 def create_activity(
     activity_in: ActivityCreate,
     session: SessionDep,
-    _: dict = Depends(verify_jwt_token),
+    current_user: Users = Depends(get_current_user),
 ):
+    activity_data = activity_in.model_dump()
+    activity_data["owner_id"] = current_user.id
     activity = Activity.model_validate(activity_in)
 
     session.add(activity)
@@ -56,12 +60,15 @@ def create_activity(
 
 @router.put("/{activity_id}", response_model=ActivityRead)
 def update_activity(
-    session: SessionDep, activity_id: int, activity_update: ActivityUpdate
+    session: SessionDep, activity_id: int, activity_update: ActivityUpdate, current_user: Users = Depends(get_current_user),
 ):
     activity = session.get(Activity, activity_id)
 
     if not activity:
         raise HTTPException(status_code=404, detail="Activity not found")
+
+    if activity.owner_id != current_user.id:
+            raise HTTPException(status_code=403, detail="Not enough permissions to edit this program")
 
     update_data = activity_update.model_dump(exclude_unset=True)
     for key, value in update_data.items():
@@ -74,10 +81,14 @@ def update_activity(
 
 
 @router.delete("/{activity_id}")
-def delete_activity(session: SessionDep, activity_id: int):
+def delete_activity(session: SessionDep, activity_id: int, current_user: Users = Depends(get_current_user),):
     activity = session.get(Activity, activity_id)
     if not activity:
         raise HTTPException(status_code=404, detail="Activity not found")
+    
+    if activity.owner_id != current_user.id:
+        raise HTTPException(status_code=403, detail="Not enough permissions to delete this program")
+
     session.delete(activity)
     session.commit()
     return {"ok": True}
